requirements.txt
google-api-python-client>=2.108.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
python-dotenv>=1.0.0
packaging>=23.2
pymysql>=1.1.0